requirements.txt
scapy
pysigma
pyyaml

sigmatools